Add test case for interface into Exporter class
Added a test case to the ExporterTest class to show the full
extent to how the method asZingChart should be used. The test
case also shows how data is to be prepared prior to calling
this method.

Updated JSON elements which required a hyphen in their id, which
could not be catered for when setting up the class names.