Implemented multi-threading support in Analyser for performance and cleaned up output from SVN command.
Added class SVNAuthor - a SVN user that performs check ins.
Added class SVNBlame - a record of the differences an SVNAuthor performs on a versioned file.
New unit tests for both SVNAuthor / SVNBlame.
New unit test for Analyser, to verify the result of analysis of a working copy folder with SVNBlame output.