Integrated both SVN and web aspects of project fully.
Added calculation of developer ratios to Analyser.
Added unit tests to verify calculation of developer ratios.
Slightly tweaked main entry point to be a more user friendly with messages.
Currently folder settings are in the app.config rather than prompting from the user, but project can now be run end-to-end in a system with SVN installed and at least one working copy folder.